# config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CHECKLIST APPLICATION",
    "version": "1.0",
    "contact": {
      "name": "Safira",
      "url": "https://github.com/safiratyas/to-do-list-backend.git",
      "email": "safeeratyas@gmail.com"
    }
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/lists/{id}/detail": {
      "put": {
        "summary": "Update",
        "description": "Change list information",
        "operationId": "updateList",
        "security": [{
          "BearerTokenAuth": []
        }],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/coreUpdateList"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [{
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "date": {
                  "type": "string",
                  "format": "date"
                }
              }
            }
          }
        ],
        "tags": [
          "List"
        ]
      }
    },
    "/api/destroys/{id}": {
      "delete": {
        "summary": "Delete List",
        "description": "Remove existing list by ID",
        "operationId": "deleteList",
        "security": [{
          "BearerTokenAuth": []
        }],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/coreListDetail"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "type": "integer",
          "format": "int64"
        }],
        "tags": [
          "List"
        ]
      }
    },
...
  },
  "definitions": {
...
    "coreUpdateList": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "date": {
          "type": "string"
        }
      }
    },
    "coreListDetail": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "rpcStatus": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array"
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerTokenAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  }
}